rename package type to cli
