+++ cli/flag_int.go
@@ -1,4 +1,4 @@
-package types
+package cli
 
 import (
 	"fmt"
