+++ cli/flag_int_test.go
@@ -1,4 +1,4 @@
-package types
+package cli
 
 import (
 	"strconv"
