+++ cli/flag_map.go
@@ -1,4 +1,4 @@
-package types
+package cli
 
 import (
 	"encoding/json"
