+++ cli/flag_string_map_deep.go
@@ -1,4 +1,4 @@
-package types
+package cli
 
 import (
 	"encoding/json"
