+++ cli/flag_string_slice_test.go
@@ -1,4 +1,4 @@
-package types
+package cli
 
 import (
 	"testing"
